fix: add Outlet to layout routes for proper nested rendering
groups.tsx, groups.$groupId.availability.tsx, and groups.$groupId.events.tsx
were rendering list content without an <Outlet>, preventing child routes
(group detail, availability request detail, event detail) from rendering.

Split each into a layout (with Outlet) and an _index route (with the list content).
